+++ src/services/PriceAggregator.ts
@@ -36,6 +36,18 @@ class PriceAggregator {
     }
   }
 
+  async valueTimestamp(symbol: string, timestamp: number): Promise<[number, number] | null> {
+    try {
+      const result = await this.connection.zrevrangebyscore(symbol, timestamp, '-inf','WITHSCORES', 'LIMIT', 0, 1);
+
+      return result.length ? [parseFloat(result[0]), parseInt(result[1])] : null;
+    } catch (err) {
+      console.error(err, JSON.stringify({symbol, timestamp}));
+
+      throw err;
+    }
+  }
+
   async averageValue(symbol: string, fromTimestamp: number, toTimestamp: number): Promise<number | null> {
     const result = await this.connection.zrevrangebyscore(symbol, toTimestamp, fromTimestamp);