+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [0.9.1] - 2021-04-08
+
+### Change
+- optimized getValidators
+- configure CryptoCompare price expiry
+- added extra logs
+
 ## [0.9.0] - 2021-04-07
 
 ### Added